// src/app/core/enums/mfe-configs.enum.ts
export enum Theme {
  DEFAULT = 'default-scheme',
  LIGHT = 'light-scheme',
  DARK = 'dark-scheme',
  BLUE = 'blue-scheme',
  DARK_BLUE = 'dark-blue-scheme',
  GREEN = 'green-scheme',
  SCM = 'scm-scheme',
}

export enum LayoutMode {
  DEFAULT = 'default',
  STICKY = 'sticky',
  TOGGLE = 'toggle',
  HIDDEN = 'hidden',
}

export enum HeaderMenuIconItem {
  HOME = 'home',
  NOTIFICATION = 'notification',
  USER_MANUALS = 'user-manuals',
  SEARCH = 'search',
  XBRAIN = 'x-brain',
}

export const PTTEP_LAYOUT_TAG = 'pttep-layout-web-component';